[SC-30890] This image is now purescript-tools, similar to haskell-tools
We now build amd64 and amr64 versions of a purescript-tools image. We
use `bookworm` as the Debian version because the `trixie` unstable version
is not currently supported by nodesource.

We use npm to install our packages. There is no arm64 version of spago,
and spago's install script doesn't fall back to building the package on
arm64, because they don't support it.

Generally, this should be fine for Flipstone's use case, but if we run
into problems, we can go back to building spago again for arm64 only.